+++ Source/Sitecore.FakeDb/DbTemplate.cs
@@ -9,7 +9,10 @@
 
     public ID ID { get; set; }
 
+    // TODO: Hide setter
     public DbFieldCollection Fields { get; set; }
+
+    internal DbFieldCollection StandardValues { get; private set; }
 
     public DbTemplate()
       : this(null)
@@ -20,6 +23,7 @@
     {
       this.Name = name;
       this.Fields = new DbFieldCollection();
+      this.StandardValues = new DbFieldCollection();
     }
 
     public DbTemplate(string name, ID id)
@@ -31,6 +35,17 @@
     public void Add(string fieldName)
     {
       this.Fields.Add(fieldName);
+    }
+
+    public void Add(string fieldName, string standardValue)
+    {
+      var id = ID.NewID;
+
+      var field = new DbField(fieldName) { ID = id };
+      this.Fields.Add(field);
+
+      var standardValueField = new DbField(fieldName) { ID = id, Value = standardValue };
+      this.StandardValues.Add(standardValueField);
     }
 
     public IEnumerator GetEnumerator()