+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 0.9.1
 * [FIX] Fixed database initialization exception appeared when adding two items with two similar set of multiple fields
+* [FIX] Simplified standard values implementation
 
 0.9.0
 * [NEW] Implemented versioned item initialization. Implemented version adding